重写了update中改变速度及伤害的方式 彻底消除了严重掉帧时 移动速度和掉血速度可能出现的问题

+++ Classes/Cannon.h
@@ -86,7 +86,7 @@ protected:
 
 	virtual void attackOnce();
 
-	int _iterator;
+	float _iterator;
 
 	virtual void onExit();
